+++ man-roxygen/args-group.R
@@ -1,2 +1,3 @@
-#' @param group A grouping variable (vector or factor). Each value in `group` is
-#'   interpreted as the group level pertaining to the corresponding observation.
+#' @param group A grouping variable. Will be coerced to [factor][base::factor]
+#'   if not already a factor. Each value in `group` is interpreted as the group
+#'   level pertaining to the corresponding observation.