+++ R/ppc-discrete.R
@@ -82,6 +82,40 @@
 #' color_scheme_set("mix-blue-pink")
 #' ppc_bars_grouped(y, yrep, group, prob = 0.5, freq = FALSE)
 #'
+#' \dontrun{
+#' # example for ordinal regression using rstanarm
+#' library(rstanarm)
+#' fit <- stan_polr(
+#'   tobgp ~ agegp,
+#'   data = esoph,
+#'   method = "probit",
+#'   prior = R2(0.2, "mean"),
+#'   init_r = 0.1,
+#'   seed = 12345,
+#'   # cores = 4,
+#'   refresh = 0
+#'  )
+#'
+#' # coded as character, so convert to integer
+#' yrep_char <- posterior_predict(fit)
+#' print(yrep_char[1, 1:4])
+#'
+#' yrep_int <- sapply(data.frame(yrep_char, stringsAsFactors = TRUE), as.integer)
+#' y_int <- as.integer(esoph$tobgp)
+#'
+#' ppc_bars(y_int, yrep_int)
+#'
+#' ppc_bars_grouped(
+#'   y = y_int,
+#'   yrep = yrep_int,
+#'   group = esoph$agegp,
+#'   freq=FALSE,
+#'   prob = 0.5,
+#'   fatten = 1,
+#'   size = 1.5
+#' )
+#' }
+#'
 NULL
 
 #' @rdname PPC-discrete
@@ -93,7 +127,7 @@ ppc_bars <-
            prob = 0.9,
            width = 0.9,
            size = 1,
-           fatten = 3,
+           fatten = 2.5,
            freq = TRUE) {
 
     dots <- list(...)
@@ -157,7 +191,7 @@ ppc_bars_grouped <-
            prob = 0.9,
            width = 0.9,
            size = 1,
-           fatten = 3,
+           fatten = 2.5,
            freq = TRUE) {
   check_ignored_arguments(...)
   call <- match.call(expand.dots = FALSE)
@@ -166,7 +200,9 @@ ppc_bars_grouped <-
   if (fixed_y(facet_args)) {
     g <- g + expand_limits(y = 1.01 * max(g$data$h))
   }
-  g + bars_group_facets(facet_args)
+  g +
+    bars_group_facets(facet_args) +
+    force_axes_in_facets()
 }