+++ NEWS.md
@@ -1,11 +1,19 @@
 <!-- See http://style.tidyverse.org/news.html for advice on writing news -->
 
-# bayesplot 1.7.0.9000
+
+<!-- 
+# bayesplot 1.7.1.9000
 
 (GitHub issue/PR numbers in parentheses)
 
-<!-- Items for next release go here-->
+Items for next release go here
+-->
 
+# bayesplot 1.7.1
+
+CRAN requested release to fix error at
+https://cran.r-project.org/web/checks/check_results_bayesplot.html due to
+matrices also inheriting from "array" in R 4.0
 
 # bayesplot 1.7.0